redisgo rename
