+++ pkg/client/redis/client_test.go
@@ -1,4 +1,4 @@
-package redisgo
+package redis
 
 import (
 	"context"
@@ -16,7 +16,7 @@ func Test_Stub(t *testing.T) {
 		var client *Client
 		defer func() {
 			if r := recover(); r != nil {
-				assert.Equal(t, r.(string), "redisgo:no master or slaves for default")
+				assert.Equal(t, r.(string), "redis:no master or slaves for default")
 				assert.Nil(t, client)
 			}
 		}()
