+++ pkg/client/redis/interceptor_test.go
@@ -1,4 +1,4 @@
-package redisgo
+package redis
 
 import (
 	"context"
